finished another dialog

+++ core/src/wrapper/RollingAverage.java
@@ -3,24 +3,24 @@ package wrapper;
 public class RollingAverage {
 
     private int size;
-    private double total = 0d;
+    private float total = 0f;
     private int index = 0;
-    private double samples[];
+    private float samples[];
 
     public RollingAverage(int size) {
         this.size = size;
-        samples = new double[size];
-        for (int i = 0; i < size; i++) samples[i] = 0d;
+        samples = new float[size];
+        for (int i = 0; i < size; i++) samples[i] = 0f;
     }
 
-    public void add(double x) {
+    public void add(float x) {
         total -= samples[index];
         samples[index] = x;
         total += x;
         if (++index == size) index = 0; // cheaper than modulus
     }
 
-    public double getAverage() {
+    public float getAverage() {
         return total / size;
     }   
 }
